Refactor proxy configuration handling: consolidate proxy file generation and introduce inline proxy configuration support
